+++ src/components/notFound/NotFound.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const NotFound: React.FC = () => {
+  return (
+    <div>
+      <h1>Page Not Found</h1>
+    </div>
+  );
+};
+
+export default NotFound;
