+++ src/pages/styles.scss
@@ -1,0 +1,52 @@
+.search-page {
+    &__header {
+        display: flex;
+        align-items: center;
+        justify-content: center;
+        height: 70px;
+    }
+
+    &__result-container {
+        position: relative;
+        width: 100%;
+        height: fit-content;
+        min-height: 500px;
+    
+        &__spinner {
+            position: absolute;
+            left: 50%;
+            top: 50%;
+            transform: translate(-50%, -50%);
+        }
+    }
+    
+    &__error-boundary {
+        background-color: #ffe6e6;
+        border: 1px solid #ff4d4d;
+        padding: 20px;
+        border-radius: 8px;
+        text-align: center;
+        color: #cc0000;
+        margin: 20px;
+    }
+    
+    &__error-btn {
+        padding: 10px 20px;
+        background-color: rgba(255, 0, 0, 0.8); // Red with 80% opacity
+        color: white;
+        border: none;
+        border-radius: 0 4px 4px 0;
+        cursor: pointer;
+        font-size: var(--font-size-m);
+        transition: background-color 0.3s ease;
+    
+        &:hover {
+            background-color: rgba(255, 0, 0, 0.9); // Less transparent on hover
+        }
+    
+        &:active {
+            background-color: rgba(255, 0, 0, 1); // Fully opaque when active
+        }
+    }
+}
+
